movi las actions del folder actions a sus correspondientes slices
