+++ client/src/store/slices/categories/index.js
@@ -1,4 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
+import axios from "axios";
 
 export const categoriesSlice = createSlice({
     name: "categories",
@@ -14,6 +15,21 @@ export const categoriesSlice = createSlice({
         },
     }
 })
+
+export function getAllCategories() {
+    return async function (dispatch) {
+        var json = await axios.get("http://localhost:3001/categories")
+        return dispatch(getCategories(json.data))
+    }
+}
+
+export function getCategoryID(id) {
+    return async function (dispatch) {
+        var json = await axios.get("http://localhost:3001/categories/" + id)
+        return dispatch(getCategoryById(json.data))
+    }
+}
+
 
 export const { getCategories, getCategoryById } = categoriesSlice.actions;
