+++ client/src/store/slices/products/index.js
@@ -1,5 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
-
+import axios from "axios";
 
 export const productsSlice = createSlice({
     name: "products",
@@ -7,16 +7,32 @@ export const productsSlice = createSlice({
         products: []
     },
     reducers: {
-        getProducts: (state, action) => {
+        setProducts: (state, action) => {
             state.products = action.payload
         },
-        getProductById: (state, action) => {
+        setProductById: (state, action) => {
             state.products = action.payload
         },
     }
 })
 
-export const { getProducts, getProductById } = productsSlice.actions;
+//funciones servicios de llamadas asincronas
+
+export function getAllProducts() {
+    return async function (dispatch) {
+        var json = await axios.get("http://localhost:3001/products")
+        return dispatch(setProducts(json.data))
+    }
+}
+
+export function getProductID(id) {
+    return async function (dispatch) {
+        var json = await axios.get("http://localhost:3001/products/" + id)
+        return dispatch(setProductById(json.data))
+    }
+}
+
+export const { setProducts, setProductById } = productsSlice.actions;
 export default productsSlice.reducer;
 
 //detail: {},
